+++ packages/cli/plugin-bff/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/plugin-bff
 
+## 3.0.5
+
+### Patch Changes
+
+- @modern-js/server-core@3.0.5
+- @modern-js/server-utils@3.0.5
+- @modern-js/builder@3.0.5
+- @modern-js/bff-core@3.0.5
+- @modern-js/create-request@3.0.5
+- @modern-js/utils@3.0.5
+
 ## 3.0.4
 
 ### Patch Changes