+++ packages/document/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @modern-js/main-doc
 
+## 3.0.5
+
+### Patch Changes
+
+- @modern-js/sandpack-react@3.0.5
+
 ## 3.0.4
 
 ### Patch Changes